Refactor JavaScript 6/11/2019
