+++ webapp/src/main/resources/static/js/functions/filter-channel.js
