+++ webapp/src/main/resources/static/js/functions/helper_functions.js
