+++ webapp/src/main/resources/static/js/functions/messageBox/updateMessages.js
@@ -1,0 +1,39 @@
+export function updateMessages() {
+    const message_box = document.getElementById("all-messages");
+    message_box.innerHTML = "";
+
+
+    const messages_promise = message_service.getAllMessagesByChannelId(channel_id);
+    messages_promise.then(messages => { //После того как Месседжи будут получены, начнется выполнение этого блока
+        messages.forEach(function (message, i) {
+            let messages_queue_context_user_container = document.createElement('div');
+            messages_queue_context_user_container.className = "c-virtual_list__item";
+            const time = message.dateCreate.split(' ')[1];
+            messages_queue_context_user_container.innerHTML = `<div class="c-message--light" id="message_${message.id}_user_${message.user.id}_content">
+                                                        <div class="c-message__gutter--feature_sonic_inputs">
+                                                            <button class="c-message__avatar__button">
+                                                                <img class="c-avatar__image">
+                                                            </button>
+                                                        </div>
+                                                        <div class="c-message__content--feature_sonic_inputs">
+                                                            <div class="c-message__content_header" id="message_${message.id}_user_${message.user.id}_content_header">
+                                                                <span class="c-message__sender">
+                                                                    <button class="c-message__sender_link">
+                                                                        ${message.user.name}
+                                                                    </button>
+                                                                </span>
+                                                                <a class="c-timestamp--static">
+                                                                    <span class="c-timestamp__label">
+                                                                        ${time}
+                                                                    </span>
+                                                                </a>
+                                                            </div>
+                                                            <span class="c-message__body">
+                                                                ${message.content}
+                                                            </span>
+                                                        </div>
+                                                    </div>`;
+            message_box.append(messages_queue_context_user_container);
+        });
+    });
+};
