+++ webapp/src/main/resources/static/js/functions/workspacePage/profileCard.js
@@ -1,0 +1,22 @@
+export const profileCard = () => {
+    // #modal_1 - селектор 1 модального окна
+    // #modal_2 - селектор 2 модального окна, которое необходимо открыть из первого
+    const two_modal = function(id_modal_1,id_modal_2) {
+        // определяет, необходимо ли при закрытии текущего модального окна открыть другое
+        let show_modal_2 = false;
+        // при нажатии на ссылку, содержащей в качестве href селектор модального окна
+        $('a[href="' + id_modal_2 + '"]').click(function(e) {
+            e.preventDefault();
+            show_modal_2 = true;
+            // скрыть текущее модальное окно
+            $(id_modal_1).modal('hide');
+        });
+        // при скрытии текущего модального окна открыть другое, если значение переменной show_modal_2 равно true
+        $(id_modal_1).on('hidden.bs.modal', function (e) {
+            if (show_modal_2) {
+                show_modal_2 = false;
+                $(id_modal_2).modal('show');
+            }
+        })
+    }('#modal_1','#modal_2');
+};
