+++ webapp/src/main/resources/static/js/functions/workspaces/updateTablesWorkspacesPage.js
@@ -1,0 +1,5 @@
+export function update_tables_workspaces_page() {
+    delete_tables_data_for_workspaces_page();
+    update_workspace_table();
+    update_add_new_workspace_table();
+};
