+++ webapp/src/main/resources/templates/workspace-page.html
@@ -11,7 +11,7 @@
     <script defer th:src="@{/js/input-box.js}" type="module"></script>
     <script defer th:src="@{/js/input-box-at.js}" type="module"></script>
     <script defer th:src="@{/js/message-box.js}" type="module"></script>
-    <script src="/js/helper_functions.js"></script>
+    <script src="/static/js/functions/helper_functions.js"></script>
     <link rel="stylesheet" th:href="@{/css/workspace-page.css}"/>
     <link rel="stylesheet" th:href="@{/css/workspace-page-old.css}"/>
     <link rel="stylesheet" th:href="@{/css/workspace-add-channel-modal.css}"/>
