+++ webapp/src/main/resources/templates/workspaces.html
@@ -7,7 +7,7 @@
     <script type="module" src="/js/workspaces.js"></script>
     <script type="module">update_workspace_table()</script>
     <script type="module">update_add_new_workspace_table()</script>
-    <script src="/js/helper_functions.js"></script>
+    <script src="/static/js/functions/helper_functions.js"></script>
 </head>
 <body>
 <table border="5" id="table_update_delete_worspaces">
